fix: set the cmake generator platform for cross-compile
