+++ src/util.ts
@@ -1,8 +1,6 @@
 import * as cp from "child_process"
 import { type PathLike, type StatOptions, type Stats, type StatsBase, stat as rawStat } from "fs-extra"
 import splitargs from "splitargs2"
-import which from "which"
-import { logger } from "./logger.js"
 
 export function getEnvVar(name: string) {
   const value = process.env[name]
@@ -10,70 +8,6 @@ export function getEnvVar(name: string) {
     return value
   }
   return undefined
-}
-
-export async function getCmakeGenerator(
-  cmake: string,
-  arch: string,
-): Promise<{
-  generator: string | undefined
-  binary: string | undefined
-}> {
-  // use ninja if available
-  const ninja = await which("ninja", { nothrow: true })
-  if (ninja !== null) {
-    return {
-      generator: "Ninja",
-      binary: ninja,
-    }
-  }
-
-  const archString = arch === "x64" ? "Win64" : arch === "x86" ? "" : null
-  if (archString === null) {
-    logger.error("Failed to find valid VS gen, using native. Good Luck.")
-    return {
-      generator: "native",
-      binary: undefined,
-    }
-  }
-
-  const generators = await execCapture(`"${cmake}" -G`)
-  const hasCR = generators.includes("\r\n")
-  const output = hasCR ? generators.split("\r\n") : generators.split("\n")
-  let found = false
-  let useVSGen = ""
-
-  for (const line of output) {
-    if (!found && line.trim() === "Generators") {
-      found = true
-      continue
-    }
-    const genParts = line.split("=")
-    if (genParts.length <= 1) {
-      // Some descriptions are multi-line
-      continue
-    }
-    /** Current MSVS compiler selected in Windows generally is prefixed with "* " */
-    genParts[0] = genParts[0].replace(/^(\* )/, "").trim()
-
-    // eslint-disable-next-line optimize-regex/optimize-regex
-    if (genParts[0].match(/Visual\s+Studio\s+\d+\s+\d+(\s+\[arch\])?/)) {
-      logger.debug("Found generator: ", genParts[0])
-      // The first entry is usually the latest entry
-      useVSGen = genParts[0]
-      break
-    }
-  }
-  const useSwitch = !useVSGen.match(/.*\[arch]/)
-  if (useSwitch) {
-    useVSGen += " -A" // essentially using this as a flag
-  } else {
-    useVSGen = useVSGen.replace("[arch]", archString).trim()
-  }
-  return {
-    generator: useVSGen,
-    binary: undefined,
-  }
 }
 
 export function execCapture(command: string): Promise<string> {
